style: 유저 정보 받는 피커 파일 분리
